requirements.txt
amqp==5.2.0
aioredis==2.0.1
appdirs==1.4.4
argh==0.26.2
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
baron==0.10.1
beautifulsoup4==4.12.3
billiard==4.2.1
black==21.6b0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
defusedxml==0.8.0rc2
dj-database-url==2.2.0
dj-email-url==1.0.6
dj-rest-auth==4.0.1
Django==5.1.3
django-allauth==0.54.0
django-auto-prefetch==1.12.0
django-appconf==1.0.6
django-autoslug==1.9.9
django-cache-url==3.4.5
django-celery-email==3.0.0
django-cors-headers==4.4.0
django-countries==7.6.1
django-environ==0.11.2
django-environs==0.1.1
django-filter==24.3
django-phonenumber-field==8.0.0
django-templated-mail==1.1.1
django-user-sessions==2.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-celery-beat==2.8.1
django-prometheus==2.4.1
djoser==2.2.3
environs==11.0.0
Faker==33.3.0
flake8==6.1.0
flower==2.0.1
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud==0.34.0
google-oauth2-tool==0.0.3
googleapis-common-protos==1.66.0
django-celery-beat==2.8.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
isort==5.12.0
incremental==24.7.2
kombu==5.4.2
marshmallow==3.21.3
msgpack==1.1.0
msgpack-python==0.5.6
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.1
phonenumbers==8.13.42
pillow==10.4.0
psycopg2==2.9.10
priority==1.3.0
prometheus_client==0.23.1
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
PyYAML==6.0.2
pyOpenSSL==24.2.1
pyparsing==3.2.0
pytest==8.3.4
pytest-factoryboy==2.7.0
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
redbaron==0.9.2
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rply==0.7.8
rsa==4.9
service-identity==24.2.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.1
tomli==2.0.2
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
wcwidth==0.2.13
watchdog==1.0.2
xhtml2pdf==0.2.17
zope.interface==7.1.1

opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-django
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-celery
